requirements.txt
openai-whisper==20231117
fsspec==2023.12.2
regex==2023.6.3
jsonschema-specifications==2023.6.1
certifi==2023.5.7
pytz==2023.3
tzdata==2023.3
pywin32==306
setuptools==65.5.0
pyzmq==25.1.0
flatbuffers==23.5.26
pip==23.3.2
attrs==23.1.0
packaging==23.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
isoduration==20.11.0
libclang==16.0.0
pyarrow==14.0.2
rich==13.7.0
more-itertools==10.2.0
Pillow==10.0.0
ipython==8.14.0
jupyter-client==8.3.0
tenacity==8.2.3
click==8.1.4
nbconvert==7.6.0
overrides==7.3.1
ipykernel==6.24.0
importlib-metadata==6.11.0
tornado==6.3.2
bleach==6.0.0
PyYAML==6.0
psutil==5.9.5
nbformat==5.9.0
traitlets==5.9.0
cachetools==5.3.1
jupyter-core==5.3.1
altair==5.2.0
chardet==5.2.0
tzlocal==5.2
decorator==5.1.1
smmap==5.0.1
tqdm==4.65.0
fonttools==4.40.0
protobuf==4.23.4
jsonschema==4.18.0
beautifulsoup4==4.12.2
rsa==4.9
opencv-python==4.8.0.74
typing-extensions==4.5.0
gitdb==4.0.11
reportlab==4.0.9
jupyterlab==4.0.2
zipp==3.17.0
filelock==3.13.1
SpeechRecognition==3.10.1
h5py==3.9.0
nltk==3.8.1
platformdirs==3.8.1
matplotlib==3.7.2
anyio==3.7.1
Markdown==3.4.3
idna==3.4
vaderSentiment==3.3.2
opt-einsum==3.3.0
oauthlib==3.2.2
networkx==3.2.1
charset-normalizer==3.2.0
GitPython==3.1.41
Jinja2==3.1.2
threadpoolctl==3.1.0
prompt-toolkit==3.0.39
pyparsing==3.0.9
audioread==3.0.1
mistune==3.0.1
markdown-it-py==3.0.0
watchdog==3.0.0
requests==2.31.0
jupyterlab-server==2.23.0
google-auth==2.21.0
pycparser==2.21
fastjsonschema==2.17.1
Pygments==2.15.1
keras==2.13.1
tensorboard==2.13.0
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
Babel==2.12.1
python-dateutil==2.8.2
jupyter-server==2.7.0
emoji==2.6.0
soupsieve==2.4.1
jsonpointer==2.4
Werkzeug==2.3.6
termcolor==2.3.0
asttokens==2.2.1
jupyter-lsp==2.2.0
MarkupSafe==2.1.3
torch==2.1.2
pywinpty==2.0.10
python-json-logger==2.0.7
async-lru==2.0.3
pandas==2.0.3
grpcio==1.56.0
streamlit==1.29.0
urllib3==1.26.16
numpy==1.24.3
six==1.16.0
cffi==1.15.1
wrapt==1.15.0
webcolors==1.13
sympy==1.12
scipy==1.11.1
Send2Trash==1.8.2
pooch==1.8.0
fpdf==1.7.2
blinker==1.7.0
debugpy==1.6.7
astunparse==1.6.3
websocket-client==1.6.1
nest-asyncio==1.5.6
fqdn==1.5.1
pandocfilters==1.5.0
kiwisolver==1.4.4
absl-py==1.4.0
ffmpeg==1.4
joblib==1.3.1
requests-oauthlib==1.3.1
mpmath==1.3.0
scikit-learn==1.3.0
sniffio==1.3.0
uri-template==1.3.0
arrow==1.2.3
tinycss2==1.2.1
executing==1.2.0
contourpy==1.1.0
msgpack==1.0.7
google-auth-oauthlib==1.0.0
numba==0.58.1
llvmlite==0.41.1
wheel==0.40.0
audiomentations==0.34.1
tensorflow-io-gcs-filesystem==0.31.0
referencing==0.29.1
pydub==0.25.1
validators==0.22.0
jedi==0.18.2
terminado==0.17.1
textblob==0.17.1
prometheus-client==0.17.0
seaborn==0.12.2
soundfile==0.12.1
toolz==0.12.0
cycler==0.11.0
toml==0.10.2
librosa==0.10.1
json5==0.9.14
rpds-py==0.8.10
parso==0.8.3
pydeck==0.8.1b0
nbclient==0.8.0
pickleshare==0.7.5
defusedxml==0.7.1
tensorboard-data-server==0.7.1
jupyter-events==0.6.3
stack-data==0.6.2
tiktoken==0.5.2
webencodings==0.5.1
pyasn1==0.5.0
colorama==0.4.6
jupyter-server-terminals==0.4.4
gast==0.4.0
soxr==0.3.7
lazy-loader==0.3
pyasn1-modules==0.3.0
PyAudio==0.2.14
wcwidth==0.2.6
notebook-shim==0.2.3
jupyterlab-pygments==0.2.2
pure-eval==0.2.2
backcall==0.2.0
google-pasta==0.2.0
matplotlib-inline==0.1.6
rfc3339-validator==0.1.4
comm==0.1.3
mdurl==0.1.2
rfc3986-validator==0.1.1
streamlit-audio-plot==0.1.0